requirements.txt
Django==1.6.7
42cc-pystyle
flake8==2.6.2
South==1.0.2
django-bootstrap-form
Pillow==2.9.0
django-smartfields==1.0.9
